Draws whatever you input in the GUI interface
