+++ src/map/Coordinate.java
@@ -26,8 +26,6 @@ public class Coordinate {
         
     }
     
-    
-
     public double getX(){
         
         return lat;
