+++ src/map/PoliticalMap.java
@@ -1,15 +1,9 @@
 /*
- * Political Map Project
- * Name: 
- * Block:
- * 
- * Program Purpose:
- *
- * Algorithm:
- * 
- * Future/possible improvements:
- *
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
  */
+
 package map;
 import edu.princeton.cs.introcs.*;
 import java.io.File;
@@ -17,26 +11,136 @@ import java.io.FileNotFoundException;
 import java.util.Scanner;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
 /**
  *
- * @author 
+ * @author hcps-mabrysh
  */
-public class PoliticalMap {
-    
-    /*public static void main(String[] args) throws Exception{
-        StdDraw.setCanvasSize(1000,665);
-        //We need to make a method for this to resize it according to each state
-        
-        File state;
+public class PoliticalMap extends javax.swing.JFrame {
+
+    /**
+     * Creates new form PoliticalMap
+     */
+    public PoliticalMap() {
+        initComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        newState = new javax.swing.JButton();
+        jLabel = new javax.swing.JLabel();
+        stateInits = new javax.swing.JTextField();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+
+        newState.setText("Enter");
+        newState.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                newStateActionPerformed(evt);
+            }
+        });
+
+        jLabel.setText("Enter State's initials:");
+
+        stateInits.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                stateInitsActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jLabel)
+                .addGap(19, 19, 19)
+                .addComponent(stateInits, javax.swing.GroupLayout.PREFERRED_SIZE, 140, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 39, Short.MAX_VALUE)
+                .addComponent(newState)
+                .addGap(35, 35, 35))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                .addContainerGap(44, Short.MAX_VALUE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(stateInits, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel)
+                    .addComponent(newState))
+                .addGap(38, 38, 38))
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void stateInitsActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_stateInitsActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_stateInitsActionPerformed
+
+    private void newStateActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_newStateActionPerformed
+        String stateName = stateInits.getText();
+        File state = new File("src\\data\\"+stateName+".txt");
         Scanner scanDoc;//scans the document
-        Scanner scanInput = new Scanner(System.in);//scans the user input
-        System.out.println("Please input the initials of the state you want to view:");
-        String stateName = scanInput.next();
-        state = new File("src\\data\\"+stateName+".txt");
-        scanDoc = new Scanner(state);       
-        InputParser input = new InputParser(stateName);
-        input.pointFinder();
+        try{
+            StdDraw.setCanvasSize(1000,665);
+            //We need to make a method for this to resize it according to each state
+            scanDoc = new Scanner(state);
+            InputParser input = new InputParser(stateName);
+            input.pointFinder();
+        }
+        catch(Exception e){
+            System.out.print("why can i not just do throws exception");
+        }
+    }//GEN-LAST:event_newStateActionPerformed
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]){
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(PoliticalMap.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(PoliticalMap.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(PoliticalMap.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(PoliticalMap.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
         
         
-    }*/
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new PoliticalMap().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JLabel jLabel;
+    private javax.swing.JButton newState;
+    private javax.swing.JTextField stateInits;
+    // End of variables declaration//GEN-END:variables
 }
