+++ src/map/Votes.java
@@ -1,0 +1,40 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+package map;
+
+/**
+ *
+ * @author hcps-taylorsd
+ */
+import java.io.*;
+import java.util.*;
+
+public class Votes {
+    
+        int[] numVotes;
+        
+        String cName;
+        
+        Votes(String c, int[] votes){
+            
+                numVotes = votes;
+                cName = c;
+            
+        }
+        
+        String splitVotes(){
+            
+            
+            
+            
+            
+            return cName;
+        }
+        
+    
+    
+}
